backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
stripe==7.4.0
redis==5.0.1
aioredis==2.0.1
email-validator==2.1.0
cors==1.0.1
python-multipart==0.0.6
cryptography==41.0.7
httpx==0.25.2
APScheduler==3.10.4
